requirements.txt
imap-tools>=1.6.0
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
schedule>=1.2.1
Pillow
pillow_heif
fastapi>=0.109.2
uvicorn>=0.27.1
pydantic>=2.6.1
PyJWT>=2.8.0
slowapi>=0.1.9
python-multipart>=0.0.9
openpyxl>=3.1.2
